Renamed MenuModal to MenuDrawer
